updated float

+++ api/models.py
@@ -4,7 +4,7 @@ class Movie(models.Model):
     name = models.CharField(max_length=50)
     description = models.TextField(max_length=360)
     year = models.IntegerField(blank=False, null=False)
-    rating = models.IntegerField(blank=False, null=False)
+    rating = models.FloatField(blank=False, null=False)
     created_date = models.DateTimeField(
         default=timezone.now)
     updated_date = models.DateTimeField(auto_now_add=True)
